requirements.txt
alembic==1.7.1
atomicwrites==1.4.0
attrs==21.2.0
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
iniconfig==1.1.1
install==1.3.4
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.13.0
packaging==21.0
pluggy==1.0.0
psycopg2==2.9.1
py==1.10.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.2.5
python-dotenv==0.19.0
SQLAlchemy==1.4.23
toml==0.10.2
webargs==8.0.1
Werkzeug==2.0.1
